refactor(): use types path
